Sources refactoring
