+++ src/tcp.h
@@ -1,0 +1,10 @@
+//
+// Created by Паша Косило on 11/3/18.
+//
+
+#ifndef PROXY_TCP_H
+#define PROXY_TCP_H
+
+#endif //PROXY_TCP_H
+
+int tcpProxy(int argc, char* argv[]);
