+++ src/udp.h
@@ -1,0 +1,10 @@
+//
+// Created by Паша Косило on 11/3/18.
+//
+
+#ifndef PROXY_UDP_H
+#define PROXY_UDP_H
+
+#endif //PROXY_UDP_H
+
+int udpProxy(int argc, char* argv[]);
